+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Types of changes
     Security in case of vulnerabilities.
 )
 
+## [2.8.6](https://github.com/pdir/maklermodul-bundle/tree/2.8.6) – 2023-03-08
+
+- [Fixed] Fix exception for list view with no objects
+- [Fixed] Fix immoscout energy pass type
+
 ## [2.8.5](https://github.com/pdir/maklermodul-bundle/tree/2.8.5) – 2023-03-06
 
 - [Fixed] Fix immoscout import